+++ app/Services/Buku/DeleteBukuService.php
@@ -4,6 +4,7 @@ namespace App\Services\Buku;
 
 use App\Base\ServiceBase;
 use App\Models\Buku;
+use App\Models\TransaksiPeminjaman;
 use App\Responses\ServiceResponse;
 use Illuminate\Support\Facades\Log;
 
@@ -18,6 +19,10 @@ class DeleteBukuService extends ServiceBase
     {
         try {
             $data = Buku::where('id', $this->id)->first();
+            $transaksi_peminjaman = TransaksiPeminjaman::where(['buku_id' => $data->id, 'tanggal_kembali' => null])->count();
+            if ($transaksi_peminjaman > 0) {
+                return self::error(null, 'Gagal menghapus data, ada buku yang masih di pinjam');
+            }
             if ($data) {
                 $data->delete();
                 return self::success(null, 'Berhasil menghapus data');
